requirements.txt
numpy~=1.22.3
tokenizers~=0.12.1
torch~=1.11.0
gin-config~=0.5.0
datasets~=2.1.0
transformers~=4.18.0
torchmetrics~=0.8.1
wandb~=0.12.15